Move GUI message dispatch into module
